+++ docs/README.md
@@ -87,7 +87,7 @@
   - 레벨5: 
 ############################################
 ```
-- 페어 매칭 결과 출력
+- 페어 매칭 결과 출력 O
 ```
 페어 매칭 결과입니다.
 이브 : 윌터
